map split functionality and support new wrapper changes
